+++ src/panoptes/data/utils/cli/main.py
@@ -42,6 +42,7 @@ def download(sequence_id: str | None = typer.Argument(..., help='Sequence ID for
 
     except Exception as e:
         print(f'[red]Error downloading images for {sequence_id}: {e}')
+        raise typer.Exit(code=1) from e
 
     return local_files
 
@@ -84,6 +85,7 @@ def get_metadata(
             print(f'[green]Metadata saved to {output_fn}')
         except Exception as e:
             print(f'[red]Error downloading metadata for {sequence_id}: {e}')
+            raise typer.Exit(code=1) from e
     else:
         if unit_id is None:
             print('[red]Must provide a unit_id if not providing a sequence_id.')
@@ -120,7 +122,8 @@ def get_metadata(
             pd.concat(dfs).to_csv(output_fn)
 
         except ValueError as e:
-            print(f'Error downloading metadata for {unit_id}: {e}', fg='red')
+            print(f'[red]Error downloading metadata for {unit_id}: {e}')
+            raise typer.Exit(code=1) from e
 
     print(f'Metadata saved to [green]{output_fn}')